+++ public/index.html
@@ -17,10 +17,31 @@
             <input type="password" placeholder="password"/>
             <button class="login-button"> Login</button>
             <p class="message">Login as Guest</p>
-            <a href class="login-button"> Guest Login </a>
+            <a id='oauth' class="login-button"> Guest Login </a>
         </form>
     </div>
 </div>
+
+<script>
+
+  let URL = "https://public-api.wordpress.com/oauth2/authorize";
+
+  let options = {
+    client_id: '70890',
+    redirect_uri: 'http://localhost:3001/oauth',
+    response_type: 'code',
+  }
+
+  let QueryString = Object.keys(options).map((key) => {
+    return `${key}=` + encodeURIComponent(options[key]);
+  }).join("&");
+
+  let authURL = `${URL}?${QueryString}`;
+
+  let link = document.getElementById('oauth');
+  link.setAttribute('href', authURL);
+
+</script>
 
 </main>
 </body>
